Add extimating global camera pose and placing 3D points in it
