+++ CMakeLists.txt
@@ -14,7 +14,7 @@ add_executable(
         "src/cameraTransition.cpp" "src/cameraTransition.h"
         "src/featureTracking.cpp" "src/featureTracking.h"
         "src/triangulate.cpp" "src/triangulate.h"
- "src/videoProcessingCycle.h" "src/videoProcessingCycle.cpp")
+ "src/videoProcessingCycle.h" "src/videoProcessingCycle.cpp" src/reportCycleForTwoFramesPair.cpp src/reportCycleForTwoFramesPair.h)
 
 target_link_libraries( slam-indoor-code ${OpenCV_LIBS} )
