+++ data/two_frames/1/3Dpoints.txt
@@ -1,4 +1,4 @@
-3D points: 943
+3D points in world system: 943
 
 [127.13669, 0.53776199, 0.13368738;
  127.05198, 1.1377702, 0.14164105;
