+++ data/two_frames/5/3Dpoints.txt
@@ -1,60 +1,60 @@
-3D points: 58
+3D points in world system: 58
 
-[48.753513, 18.375841, 1.0245023;
- 52.696735, 3.707777, 0.75967097;
- 50.228668, 15.384864, 0.89998239;
- 52.625065, 4.0305161, 0.76074994;
- 52.536274, 4.5436616, 0.76210934;
- 52.361267, 5.2399907, 0.76196772;
- 52.383919, 5.2506256, 0.76225531;
- 52.671719, 3.9232278, 0.74297637;
- 52.139961, 6.2395062, 0.76461631;
- 52.116203, 6.3481359, 0.76473856;
- 51.531048, 9.2563848, 0.77941078;
- 51.461884, 9.5918808, 0.77867222;
- 51.59584, 9.1263962, 0.77625561;
- 51.31633, 10.287749, 0.77741361;
- 50.889561, 12.326224, 0.77258301;
- 50.623951, 13.643112, 0.74642533;
- 51.951405, 7.4930711, 0.7251659;
- 50.858097, 12.657964, 0.74276149;
- 51.911854, 7.6603489, 0.72451663;
- 50.314983, 15.122629, 0.72102123;
- 50.356403, 14.939523, 0.72060859;
- 50.360123, 14.924603, 0.71993423;
- 43.870129, 45.026665, 0.73836261;
- 49.879112, 17.213352, 0.70825863;
- 49.866814, 17.292631, 0.70807952;
- 49.774349, 17.780476, 0.70796001;
- 46.630424, 32.354691, 0.71128118;
- 48.59343, 23.22942, 0.70475864;
- 39.167873, 67.694336, 0.71254569;
- 39.331272, 66.944191, 0.71186131;
- 48.863514, 22.074896, 0.70186281;
- 48.885056, 21.876955, 0.70150894;
- 30.075811, 110.99293, 0.71468878;
- 48.863312, 21.979864, 0.70126069;
- 48.862717, 21.983017, 0.70117432;
- 48.906662, 21.779118, 0.70072472;
- 48.957943, 21.581776, 0.70043612;
- 16.413372, 175.55907, 0.71706653;
- 49.014179, 21.278198, 0.70010507;
- 49.002487, 21.336296, 0.69932157;
- 49.08136, 21.005728, 0.69883728;
- 49.016712, 21.352613, 0.69875938;
- 49.121563, 20.853828, 0.69866955;
- 49.131622, 20.806442, 0.69866878;
- 40.136345, 63.099567, 0.68918633;
- 41.399853, 56.846592, 0.69045228;
- 48.597061, 23.303528, 0.6934877;
- 31.027103, 105.52496, 0.67605931;
- 32.745922, 98.124359, 0.67123103;
- 37.365589, 76.252495, 0.67281765;
- 41.055637, 58.844654, 0.67849481;
- 37.360523, 75.907684, 0.67307931;
- 14.219893, 184.26913, 0.63777655;
- 46.203693, 34.513504, 0.68643671;
- 10.419429, 202.06935, 0.62651294;
- 37.981071, 73.244225, 0.67202914;
- 47.797356, 27.069485, 0.68780899;
- 30.902605, 106.80312, 0.65187138]
+[49.216656, 19.706198, 0.52000499;
+ 53.159878, 5.0381341, 0.25517368;
+ 50.691811, 16.715221, 0.3954851;
+ 53.088207, 5.3608732, 0.25625265;
+ 52.999416, 5.8740187, 0.25761205;
+ 52.824409, 6.5703478, 0.25747043;
+ 52.847061, 6.5809827, 0.25775802;
+ 53.134861, 5.2535849, 0.23847908;
+ 52.603104, 7.5698633, 0.26011902;
+ 52.579346, 7.678493, 0.26024127;
+ 51.99419, 10.586742, 0.27491349;
+ 51.925026, 10.922238, 0.27417493;
+ 52.058983, 10.456754, 0.27175832;
+ 51.779472, 11.618107, 0.27291632;
+ 51.352703, 13.656582, 0.26808572;
+ 51.087093, 14.97347, 0.24192804;
+ 52.414547, 8.8234282, 0.22066861;
+ 51.321239, 13.988321, 0.2382642;
+ 52.374996, 8.9907064, 0.22001934;
+ 50.778126, 16.452986, 0.21652395;
+ 50.819546, 16.26988, 0.2161113;
+ 50.823265, 16.254961, 0.21543694;
+ 44.333271, 46.357021, 0.23386532;
+ 50.342255, 18.543709, 0.20376134;
+ 50.329956, 18.622988, 0.20358223;
+ 50.237492, 19.110832, 0.20346272;
+ 47.093567, 33.685047, 0.20678389;
+ 49.056572, 24.559776, 0.20026135;
+ 39.631016, 69.024696, 0.2080484;
+ 39.794415, 68.274551, 0.20736402;
+ 49.326656, 23.405252, 0.19736552;
+ 49.348198, 23.207312, 0.19701165;
+ 30.538956, 112.32329, 0.21019149;
+ 49.326454, 23.310221, 0.1967634;
+ 49.325859, 23.313374, 0.19667703;
+ 49.369804, 23.109474, 0.19622743;
+ 49.421085, 22.912132, 0.19593883;
+ 16.876516, 176.88942, 0.21256924;
+ 49.477322, 22.608555, 0.19560778;
+ 49.46563, 22.666653, 0.19482428;
+ 49.544502, 22.336084, 0.19433999;
+ 49.479855, 22.68297, 0.19426209;
+ 49.584705, 22.184185, 0.19417226;
+ 49.594765, 22.136799, 0.19417149;
+ 40.599487, 64.429924, 0.18468904;
+ 41.862995, 58.176949, 0.18595499;
+ 49.060204, 24.633884, 0.18899041;
+ 31.490248, 106.85532, 0.17156202;
+ 33.209064, 99.45472, 0.16673374;
+ 37.828732, 77.582855, 0.16832036;
+ 41.51878, 60.175011, 0.17399752;
+ 37.823666, 77.238045, 0.16858202;
+ 14.683037, 185.59949, 0.13327926;
+ 46.666836, 35.843861, 0.18193942;
+ 10.882573, 203.3997, 0.12201566;
+ 38.444214, 74.574585, 0.16753185;
+ 48.260498, 28.399841, 0.1833117;
+ 31.365749, 108.13348, 0.14737409]
