+++ data/two_frames/6/3Dpoints.txt
@@ -1,69 +1,69 @@
-3D points: 67
+3D points in world system: 67
 
-[6.4385309, 0.61676991, -0.19670856;
- 6.3009009, 0.88163668, -0.20172042;
- 6.109374, 1.2522221, -0.20384364;
- 6.1051726, 1.2522861, -0.20405373;
- 5.2722445, 2.8540442, -0.21015041;
- 5.9782438, 1.511755, -0.20596954;
- 5.24263, 2.941638, -0.2115697;
- 5.2774954, 2.8565195, -0.21123549;
- 5.2348394, 2.9558828, -0.21170729;
- 5.8824282, 1.6961217, -0.20847076;
- 5.9057856, 1.6503741, -0.2082843;
- 5.9213572, 1.6183444, -0.20818157;
- 5.9228201, 1.6169628, -0.20816945;
- 4.8804898, 3.6229146, -0.2187169;
- 4.8885245, 3.6074195, -0.21863969;
- 4.7946925, 3.7892001, -0.21985713;
- 4.8031821, 3.7728229, -0.21977422;
- 5.4160771, 2.5895946, -0.21385896;
- 5.1864886, 3.0391173, -0.21666551;
- 4.887383, 3.6186328, -0.2202702;
- 4.8825288, 3.6229641, -0.22040246;
- 4.930593, 3.5304091, -0.2200707;
- 4.8387566, 3.7084472, -0.22128445;
- 4.7962918, 3.7907047, -0.22181155;
- 4.7259526, 3.9270952, -0.2227505;
- 4.7895436, 3.8046024, -0.22230582;
- 4.8028054, 3.7797742, -0.22260162;
- 4.807622, 3.7769647, -0.22256723;
- 4.539124, 4.2969465, -0.22568089;
- 4.331985, 4.694488, -0.22801492;
- 4.2719889, 4.8109632, -0.22887912;
- 4.2339835, 4.8846517, -0.22935902;
- 3.3863153, 6.5519924, -0.23883089;
- 4.1508331, 5.0466251, -0.23035939;
- 4.7993078, 3.7872906, -0.22306314;
- 4.2180138, 4.9157348, -0.22964221;
- 4.074203, 5.1959977, -0.23128897;
- 3.3052304, 6.6932087, -0.24001585;
- 3.7874684, 5.7554135, -0.23466454;
- 4.2732806, 4.8089571, -0.22921266;
- 4.6838784, 4.0118394, -0.22464725;
- 4.2488079, 4.8615208, -0.22963467;
- 4.0959282, 5.1467071, -0.23157993;
- 4.5974169, 4.1848412, -0.22588176;
- 2.4494319, 8.3503933, -0.2506007;
- 4.2856078, 4.7986379, -0.22954485;
- 4.490819, 4.3911915, -0.22735548;
- 3.8912125, 5.5693917, -0.23437791;
- 4.706203, 3.9743056, -0.22499578;
- 5.1597004, 3.0921824, -0.21973254;
- 3.6644304, 6.0039101, -0.23746754;
- 3.7173572, 5.899971, -0.23691997;
- 3.7268081, 5.8815527, -0.23683789;
- 3.8133733, 5.707819, -0.23586638;
- 4.027667, 5.2914486, -0.23351188;
- 4.0821972, 5.1855187, -0.23292451;
- 4.0742235, 5.2010174, -0.23305568;
- 4.1800685, 4.9954004, -0.23182796;
- 4.1561003, 5.0419674, -0.23213717;
- 4.2143044, 4.9289069, -0.23147534;
- 4.7816162, 3.8273566, -0.22467197;
- 4.2039933, 4.9489393, -0.23160981;
- 4.2270899, 4.9040856, -0.2313883;
- 4.2582793, 4.8473926, -0.23101588;
- 4.2631264, 4.8379602, -0.23096852;
- 4.4149475, 4.5392747, -0.2292988;
- 4.4558654, 4.459815, -0.22883378]
+[7.837821, 2.1079645, -0.38853654;
+ 7.700191, 2.3728313, -0.3935484;
+ 7.5086641, 2.7434168, -0.39567164;
+ 7.5044627, 2.7434807, -0.39588171;
+ 6.6715345, 4.3452387, -0.4019784;
+ 7.3775339, 3.0029497, -0.39779752;
+ 6.6419201, 4.4328327, -0.40339768;
+ 6.6767855, 4.3477139, -0.40306348;
+ 6.6341295, 4.4470773, -0.40353528;
+ 7.2817183, 3.1873164, -0.40029874;
+ 7.3050756, 3.1415687, -0.4001123;
+ 7.3206472, 3.109539, -0.40000957;
+ 7.3221102, 3.1081574, -0.39999744;
+ 6.2797799, 5.114109, -0.4105449;
+ 6.2878146, 5.0986142, -0.41046768;
+ 6.1939826, 5.2803946, -0.41168511;
+ 6.2024722, 5.2640176, -0.4116022;
+ 6.8153672, 4.0807891, -0.40568694;
+ 6.5857787, 4.5303121, -0.40849349;
+ 6.2866731, 5.1098275, -0.4120982;
+ 6.2818189, 5.1141586, -0.41223046;
+ 6.3298831, 5.0216036, -0.4118987;
+ 6.2380466, 5.1996417, -0.41311243;
+ 6.1955819, 5.2818995, -0.41363955;
+ 6.1252427, 5.4182897, -0.4145785;
+ 6.1888337, 5.2957969, -0.41413382;
+ 6.2020955, 5.2709689, -0.41442961;
+ 6.206912, 5.2681594, -0.41439521;
+ 5.9384141, 5.7881413, -0.41750887;
+ 5.7312751, 6.1856828, -0.4198429;
+ 5.671279, 6.3021579, -0.42070711;
+ 5.6332736, 6.3758464, -0.42118701;
+ 4.7856054, 8.0431871, -0.43065888;
+ 5.5501232, 6.5378199, -0.42218739;
+ 6.1985979, 5.2784853, -0.41489112;
+ 5.6173038, 6.4069295, -0.4214702;
+ 5.4734931, 6.6871924, -0.42311695;
+ 4.7045202, 8.1844034, -0.43184385;
+ 5.1867585, 7.2466083, -0.42649254;
+ 5.6725707, 6.3001518, -0.42104065;
+ 6.0831685, 5.5030341, -0.41647524;
+ 5.648098, 6.3527155, -0.42146266;
+ 5.4952183, 6.6379018, -0.42340791;
+ 5.996707, 5.6760359, -0.41770974;
+ 3.848722, 9.841588, -0.44242868;
+ 5.6848979, 6.2898326, -0.42137283;
+ 5.8901091, 5.8823862, -0.41918346;
+ 5.2905025, 7.0605865, -0.4262059;
+ 6.1054931, 5.4655004, -0.41682377;
+ 6.5589905, 4.5833769, -0.41156054;
+ 5.0637202, 7.4951048, -0.42929554;
+ 5.1166472, 7.3911657, -0.42874795;
+ 5.1260982, 7.3727474, -0.42866588;
+ 5.2126632, 7.1990137, -0.42769438;
+ 5.4269571, 6.7826433, -0.42533988;
+ 5.4814873, 6.6767135, -0.4247525;
+ 5.4735136, 6.6922121, -0.42488366;
+ 5.5793586, 6.4865952, -0.42365596;
+ 5.5553904, 6.5331621, -0.42396516;
+ 5.6135945, 6.4201016, -0.42330334;
+ 6.1809063, 5.3185511, -0.41649997;
+ 5.6032834, 6.440134, -0.4234378;
+ 5.62638, 6.3952804, -0.42321628;
+ 5.6575694, 6.3385873, -0.42284387;
+ 5.6624165, 6.329155, -0.42279652;
+ 5.8142376, 6.0304694, -0.42112678;
+ 5.8551555, 5.9510098, -0.42066178]
